✨feat: 로그인 후 비밀변호 변경

+++ woorepie/src/main/java/com/piehouse/woorepie/customer/dto/response/GetCustomerAccountResponse.java
@@ -23,6 +23,6 @@ public class GetCustomerAccountResponse {
 
     private Integer estateTokenPrice;
 
-    private BigDecimal estatePrice;
+    private Integer estatePrice;
 
 }
